+++ public/css/components/copyright-policy.css
@@ -1,0 +1,201 @@
+.breadcrumb {
+    padding: 20px 0;
+    font-size: 0.9rem;
+    color: #666;
+    margin-bottom: 20px;
+}
+
+.breadcrumb a {
+    background: var(--gradient-text-main);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    text-decoration: none;
+    transition: all 0.3s ease;
+    font-weight: 500;
+}
+
+.breadcrumb a:hover {
+    background: var(--gradient-horizontal);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+}
+
+.copyright-policy {
+    display: flex;
+    flex-direction: column;
+    min-height: 100vh;
+}
+
+.copyright-main {
+    display: grid;
+    grid-template-columns: 200px 1fr 200px;
+    gap: 40px;
+    max-width: 1400px;
+    margin: 0 auto;
+    padding: 40px 20px;
+    min-height: calc(100vh - 200px);
+    align-items: start;
+}
+
+.terms-wrapper {
+    max-width: 900px;
+    margin: 0 auto;
+    padding: 40px 32px;
+    background: rgba(255, 255, 255, 0.95);
+}
+
+.poster-strip {
+    position: sticky;
+    top: 100px;
+    width: 200px;
+    height: calc(100vh - 200px);
+    max-height: 1000px;
+    overflow: hidden;
+    background: linear-gradient(135deg, rgba(0,0,0,0.8) 0%, rgba(0,0,0,0.4) 100%);
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.poster-container {
+    position: relative;
+    width: 100%;
+    height: 100%;
+    overflow: hidden;
+}
+
+.poster-container img {
+    position: absolute;
+    top: 0;
+    left: 0;
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    opacity: 0;
+    transition: all 1s ease-in-out;
+    box-shadow: 0 8px 25px rgba(0, 0, 0, 0.6);
+}
+
+.poster-container img.active {
+    opacity: 1;
+}
+
+.poster-container img:hover {
+    transform: scale(1.05);
+    box-shadow: 0 15px 40px rgba(229, 157, 44, 0.5);
+    z-index: 10;
+}
+
+.poster-strip::before {
+    content: '';
+    position: absolute;
+    top: -2px;
+    left: -2px;
+    right: -2px;
+    bottom: -2px;
+    background: var(--gradient-main);
+    border-radius: 17px;
+    z-index: -1;
+    opacity: 0.3;
+    animation: glow 3s ease-in-out infinite alternate;
+}
+
+.terms-wrapper h1 {
+    text-align: center;
+    font-size: 2.6rem;
+    margin-bottom: 40px;
+    background: var(--gradient-text-main);
+    -webkit-background-clip: text;
+    -webkit-text-fill-color: transparent;
+    font-weight: 700;
+    position: relative;
+}
+
+.terms-wrapper h1::after {
+    content: '';
+    position: absolute;
+    bottom: -10px;
+    left: 50%;
+    transform: translateX(-50%);
+    width: 80px;
+    height: 3px;
+    background: var(--gradient-horizontal);
+    border-radius: 2px;
+}
+
+.terms-wrapper h2 {
+    font-size: 1.6rem;
+    margin: 40px 0 20px;
+    background: var(--gradient-horizontal);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    border-left: 4px solid #e59d2c;
+    padding-left: 16px;
+    font-weight: 600;
+}
+
+.terms-wrapper h3 {
+    font-size: 1.3rem;
+    margin: 30px 0 15px;
+    background: var(--gradient-horizontal);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    font-weight: 600;
+}
+
+.terms-wrapper p {
+    line-height: 1.8;
+    text-align: justify;
+    color: #333;
+    margin-bottom: 16px;
+    font-size: 1rem;
+}
+
+.terms-wrapper ul {
+    margin: 16px 0;
+    padding-left: 20px;
+}
+
+.terms-wrapper li {
+    line-height: 1.7;
+    margin-bottom: 8px;
+    color: #333;
+}
+
+.terms-wrapper strong {
+    background: var(--gradient-horizontal);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    font-weight: 600;
+}
+
+.terms-wrapper em {
+    background: var(--gradient-horizontal);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    font-style: italic;
+}
+
+.terms-wrapper a {
+    background: var(--gradient-horizontal);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+    text-decoration: none;
+    font-weight: 500;
+    transition: all 0.3s ease;
+}
+
+.terms-wrapper a:hover {
+    text-decoration: underline;
+    background: var(--gradient-main);
+    -webkit-background-clip: text;
+    background-clip: text;
+    -webkit-text-fill-color: transparent;
+}
