+++ public/css/components/modal.css
@@ -39,7 +39,7 @@
 
 .modal-content {
   background-color: #000;
-  margin: 9% auto;
+  margin: 5% auto;
   padding: 25px;
   border: 10px solid transparent;
   background: linear-gradient(#000, #000) padding-box,
@@ -47,6 +47,14 @@
   max-width: 1000px;
   min-height: 50vh;
   animation: slideIn 0.3s ease;
+  position: relative;
+}
+
+#youtubeVideo {
+  width: 100%;
+  height: 400px;
+  border: none;
+  border-radius: 8px;
 }
 
 .close,
