+++ public/css/style.css
@@ -29,3 +29,4 @@
 @import url('components/invoice.css');
 @import url('components/showtime-management.css');
 @import url('components/auth-profile.css');
+@import url('components/copyright-policy.css');
