+++ public/js/modules/copyright-policy.js
@@ -1,0 +1,6 @@
+import { initCopyrightPolicyPosters } from './policy-poster-strips.js';
+
+export function initCopyrightPolicy() {
+    console.log('©️ Khởi tạo Copyright Policy page...');
+    return initCopyrightPolicyPosters();
+}
